### 2020-03-04 #20 @khinsen
[¬Rp] Stokes drag on conglomerates of spheres
**Original article:** [B. Cichocki, K. Hinsen, Stokes drag on conglomerates of spheres, Phys. Fluids 7:286 (1995)](https://doi.org/10.1063/1.868626)

**Scientific domain:**  Fluid dynamics
**Programming language:**  Fortran 77

Since this is a failed reproduction attempt, no results are presented, meaning that editors and reviewers do not need to be familiar with the application domain .
