### 2020-05-13 #20 @otizonaizit
Hi @rougier , @khinsen : is this a submission connected to the Ten Years Reproducibility Challenge? It seems hard to get reviewers...
