### 2020-05-13 #20 @khinsen
@otizonaizit Yes, it's a contribution to the challenge. And since it's a failed reproduction attempt, it doesn't require any domain expertise (in colloid science) to evaluate.
