### 2020-05-28 khinsen/rescience-ten-year-challenge-paper-4#1 @rth
Few issues when following instructions in notes.org 
When running instruction from `notes.org` as part of the review of https://github.com/ReScience/submissions/issues/20#issuecomment-630857032, I run into a few issues. Mostly likely because I'm not familiar with Guix, running those instruction on Debian 10 without Guix does produce the expected final result.

I'm running instructions inside [nextjournal/guix](https://hub.docker.com/r/nextjournal/guix/tags) Docker image based on Alpine Linux as I had some trouble installing Guix otherwise (and didn't want spend too much time debugging it)
```
# Note: guix fails to install packages without --privileged flag for some reason
docker run -it --rm -v /absolute-path/rescience-ten-year-challenge-paper-4:/project/ --privileged nextjournal/guix:1.0.1
```

The instructions do work in the end and most of the following issues are likely due to Alpine Linux including only the bare minimum of dependencies so a few of the missing ones (namely `binutils` and `glibc`) had to be added to the list of dependencies.

### 1. gfortran-toolchain: unknown package
```
guix environment --pure --ad-hoc tcsh coreutils make sed grep gfortran-toolchain -- tcsh install_lib
```
produces the following error
```
accepted connection from pid 1031, user root
guix environment: error: gfortran-toolchain: unknown package
```

Replacing `gfortran-toolchain` with `gfortran` does work.
```
guix --version
guix (GNU Guix) 1.0.1
```
### 2. ar: Command not found

Running,
```
guix environment --pure --ad-hoc tcsh coreutils make sed grep gfortran -- tcsh install_lib
```
inside `HYDROLIB` produces,
```
...
make: ar: Command not found
make: *** [Makefile:51: libblas.a] Error 127
```
Adding `binutils` to the above list does solve this issue.

### 3. cannot find crt1.o: No such file or directory
Running,
```
guix environment --pure --ad-hoc tcsh coreutils sed gawk grep gfortran binutils
./fc example.f
./a.out
```
produces,
```
/fc example.f
/gnu/store/sz9d20kmspf3caj0v6r37g9vcia5xj46-profile/bin/ld: cannot find crt1.o: No such file or directory
```
adding `glibc` to the list of dependencies does fix it.

### 4. Other comments

Just a comment that I had to replace each `guix_command` with `giux` which makes copy-pasting harder. Are there setups where `guix` does not exist but `guix_command` does?

Also I find a bit confusing, likely because of my lack of familiarity with Giux, that `guix environment` with multi-line commands seems to create a new environment in which for instance `guix` is not in the PATH, so I had to do `Ctrl+D` after each such command to return to the normal shell.

If it helps, running,
```
guix environment --container --ad-hoc guix guile -- guile -s show-dependencies.scm wget unzip tcsh coreutils sed gawk grep m
ake gfortran
```
for me seems to produce a slightly different set of dependencies,
```
accepted connection from pid 3952, user root
;;; note: auto-compilation is enabled, set GUILE_AUTO_COMPILE=0
;;;       or pass the --no-auto-compile argument to disable.
;;; compiling /project/show-dependencies.scm
;;; compiled /root/.cache/guile/ccache/2.2-LE-8-3.A/project/show-dependencies.scm.go
Packages: 9
  coreutils@8.30 gawk@4.2.1 gfortran@8.3.0 grep@3.1 make@4.2.1 sed@4.5 tcsh@6.20.00 unzip@6.0 wget@1.20.1
Package inputs: 24 packages
  acl@2.2.52 autoconf@2.69 bzip2@1.0.6 gmp@6.1.2 gnutls@3.6.5 guile@2.0.14 isl@0.19 libcap@2.25 libelf@0.8.13 libidn2@2.0.5 libpsl@0.20.2 libsigsegv@2.12 lzip@1.20 mpc@1.1.0 mpfr@4.0.1 ncurses@6.1 pcre@8.42 perl-http-daemon@6.01 perl-io-socket-ssl@2.038 perl@5.28.0 pkg-config@0.29.2 python@3.7.0 texinfo@6.5 zlib@1.2.11
Build inputs: 51 packages
  [source code from (ftp://ftp.astron.com/pub/tcsh/tcsh-6.20.00.tar.gz ftp://ftp.astron.com/pub/tcsh/old/tcsh-6.20.00.tar.gz)] [source code from mirror://gnu/coreutils/coreutils-8.30.tar.xz] [source code from mirror://gnu/gawk/gawk-4.2.1.tar.xz] [source code from mirror://gnu/gcc/gcc-8.3.0/gcc-8.3.0.tar.xz] [source code from mirror://gnu/grep/grep-3.1.tar.xz] [source code from mirror://gnu/make/make-4.2.1.tar.bz2] [source code from mirror://gnu/sed/sed-4.5.tar.xz] [source code from mirror://gnu/wget/wget-1.20.1.tar.lz] [source code from mirror://sourceforge/infozip/UnZip%206.x%20%28latest%29/UnZip%206.0/unzip60.tar.gz] acl@2.2.52 autoconf@2.69 bash-minimal@4.4.23 binutils@2.31.1 bzip2@1.0.6 coreutils@8.30 diffutils@3.6 file@5.33 findutils@4.6.0 gawk@4.2.1 gcc@5.5.0 glibc-utf8-locales@2.28 glibc@2.28 gmp@6.1.2 gnutls@3.6.5 grep@3.1 guile@2.0.14 gzip@1.9 isl@0.19 ld-wrapper@0 libcap@2.25 libelf@0.8.13 libidn2@2.0.5 libpsl@0.20.2 libsigsegv@2.12 lzip@1.20 make@4.2.1 mpc@1.1.0 mpfr@4.0.1 ncurses@6.1 patch@2.7.6 pcre@8.42 perl-http-daemon@6.01 perl-io-socket-ssl@2.038 perl@5.28.0 pkg-config@0.29.2 python@3.7.0 sed@4.5 tar@1.30 texinfo@6.5 xz@5.2.4 zlib@1.2.11
Package closure: 126 packages
  acl@2.2.52 attr@2.4.47 autoconf@2.69 bash-minimal@4.4.23 bash-static@4.4.23 bash@4.4.23 binutils-bootstrap@0 binutils-cross-boot0@2.31.1 binutils@2.31.1 bison@3.0.5 bootstrap-binaries@0 bzip2@1.0.6 coreutils@8.30 diffutils-boot0@3.6 diffutils@3.6 ed@1.14.2 expat@2.2.6 file-boot0@5.33 file@5.33 findutils-boot0@4.6.0 findutils@4.6.0 fontconfig@2.13.1 freetype@2.9.1 gawk@4.2.1 gcc-bootstrap@0 gcc-cross-boot0-wrapped@5.5.0 gcc-cross-boot0@5.5.0 gcc@5.5.0 gdbm@1.18 gettext-boot0@0.19.8.1 gettext-minimal@0.19.8.1 gfortran@8.3.0 glibc-bootstrap@0 glibc-intermediate@2.28 glibc-utf8-locales@2.28 glibc@2.28 gmp@6.1.2 gnutls@3.6.5 gperf@3.1 grep@3.1 gs-fonts@8.11 guile-bootstrap@2.0 guile@2.0.14 guile@2.2.4 gzip@1.9 isl@0.19 ld-wrapper-boot0@0 ld-wrapper-boot3@0 ld-wrapper@0 libatomic-ops@7.6.6 libbsd@0.9.1 libcap@2.25 libelf@0.8.13 libffi@3.2.1 libgc@7.6.6 libgcrypt@1.8.3 libgpg-error@1.32 libidn2@2.0.5 libltdl@2.4.6 libpng@1.6.34 libpsl@0.20.2 libpthread-stubs@0.4 libsigsegv@2.12 libstdc++-boot0@4.9.4 libstdc++@5.5.0 libtasn1@4.13 libunistring@0.9.10 libx11@1.6.6 libxau@1.0.8 libxcb@1.13 libxdmcp@1.1.2 libxext@1.3.3 libxft@2.3.2 libxml2@2.9.8 libxrender@0.9.10 libxslt@1.1.32 linux-libre-headers@4.14.67 lzip@1.20 m4@1.4.18 make-boot0@4.2.1 make@4.2.1 mpc@1.1.0 mpfr@4.0.1 ncurses@6.1 net-base@5.3 net-tools@1.60-0.479bb4a nettle@3.4.1 openssl@1.0.2p patch@2.7.6 pcre@8.42 perl-boot0@5.28.0 perl-encode-locale@1.05 perl-http-daemon@6.01 perl-http-date@6.02 perl-http-message@6.15 perl-io-html@1.00 perl-io-socket-ssl@2.038 perl-lwp-mediatypes@6.02 perl-net-ssleay@1.85 perl-test-needs@0.002005 perl-try-tiny@0.22 perl-uri@1.73 perl@5.28.0 pkg-config@0.29.2 python-minimal-wrapper@3.7.0 python-minimal@3.7.0 python2@2.7.15 python@3.7.0 readline@7.0.5 sed@4.5 sqlite@3.24.0 tar@1.30 tcl@8.6.8 tcsh@6.20.00 texinfo@6.5 tk@8.6.8 unzip@6.0 util-linux@2.32.1 util-macros@1.19.2 wget@1.20.1 which@2.21 xcb-proto@1.13 xorgproto@2018.4 xtrans@1.3.5 xz@5.2.4 zlib@1.2.11
```
