### 2020-05-28 #20 @rth
A quick status update: I have read both papers and run the code, I'll try to finish the review this weekend.
