### 2020-05-29 khinsen/rescience-ten-year-challenge-paper-4#1 @khinsen
Thanks @rth for your report!

The missing `gfortran-toolchain` is easy to explain: Guix 1.0.1 was released in May 2019, whereas `gfortran-toolchain` was added in December 2019. In fact, I submitted this addition myself in the course of working on my reproduction attempt. As you discovered experimentally, `gfortran-toolchain` is a combination of `gfortran`, `glibc`, and `binutils`, packaged in such a way that the compiled binaries contain implicit references to `glibc` to make them usable in environments that don't contain `glibc` explicitly.Y our points 1 to 3 are all covered by  `gfortran-toolchain`.

I will add the date of the Guix version I used to my notes so that readers are warned. I do list the commit, but that isn't easy to translate into a point in the timeline.

As for `guix_command`, it is actually defined in my `notes.org`, but you have to look at the raw text because GitHub's formatting of org-mode files make it invisible:
```
#+NAME: guix-command
guix time-machine --commit=f76c16d220e6c349441c08bf25a5197037490fa5 --
```
It is then transformed into a shell variable in every code block header:
```
#+begin_src sh :results output :exports both :var guix_command=guix-command
```
The point of this is to make continuous integration testing easier, because all it takes is changing the Guix commit reference.

Substituting `guix_command` command for plain `guix` comes down to using [commit f76c16d220e6c349441c08bf25a5197037490fa5](https://archive.softwareheritage.org/browse/revision/f76c16d220e6c349441c08bf25a5197037490fa5/), dated 28 February 2020, which is what I used for my work. That's what ensures bit-level reproducibility.

Unfortunately, with Guix 1.0.1,  this commit-specific command won't work either because `guix time-machine` was added only in November 2019. That's also something I will add to my notes. It is interesting to note, however, that `guix time-machine` does work for the future. You can install a Guix from early December and then install the `gfortran-toolchain` package, even though it was added later!
