### 2020-05-29 e46f158
Explanation about the minimal Guix versions required to run the code
Fixes #1
