### 2020-05-31 khinsen/rescience-ten-year-challenge-paper-4#1 @rth
Thanks for the explanations! I can confirm it's working as expected with a later guix version when first applying `guix time-machine`

> Substituting guix_command command for plain guix comes down to using commit f76c16d220e6c349441c08bf25a5197037490fa5, dated 28 February 2020, which is what I used for my work. That's what ensures bit-level reproducibility.

I can use the `guix time-machine` but that still didn't define the `guix-command` for me. The,
```
#+NAME: guix-command
[...]
#+begin_src sh :results output :exports both :var guix_command=guix-command
```
is something tied to Org-Mode and Emacs as far as I understand? I ended up doing an
```
alias guix-command=guix
```
