### 2020-05-31 #20 @rth
### General comments

This paper describes a failed reproduction attempt for computing stokes drag on conglomerates of spheres with a Fortran code.
The two takeaways for me were that:
 - that archival of the source code is not enough, as missing configuration files, which are often considered less important, can also prevent reproducibility.
 - that the Fortran code in question still works as expected 34 years later with mostly no changes, which cannot be said about more modern programming tools.

Overall I think this paper presents a nice prospective on code reproducibility and provides an interesting example of code life cycle for code used in research over 34 years.

### Code

I was able to run the code following instructions in the code repository, and obtained identical results to those presented. Some technical questions were answered in https://github.com/khinsen/rescience-ten-year-challenge-paper-4/issues/1
The use of Guix for code reproducibility is very helpful.

A few minor comments:
 - I think that the use of org-mode notes can be confusing, as some of code comments (that are not rendered on Github) have an impact on environment variables. I does work in the end, bit it's just one more thing the person following these instructions needs to be aware of.
 - It might have been easier to create one Guix environment with all required dependencies, then run all shell commands there. As opposed to running `guix environment` multiple times with different dependencies and having to explicitly exit this environment after each step. At least that was the case for me, maybe I was doing something wrong due to my lack of familiarity with Guix. The use or not of org-mode may be the reason behind it as well.
 - Just a note that the source code for the Fortran code, currently on Github, can also be accessed [via the Software Heritage archive](https://archive.softwareheritage.org/browse/origin/directory/?origin_url=https://github.com/khinsen/HYDROLIB)
 - There is likely a typo on the sentence,
   > The column “final” corresponds to L=4 for the smaller conglomerates (N=55 or less) and to L=4 for the larger ones.

   in `summary.org` file.

### Paper

It's an interesting read and I don't have any major comments on the form or the contents.

One comment is that upon reading "reproduction failure" the first though is that the original paper had issues or that the code has portability limitations. It's not the case here (aside from the configuration files needed to reproduce not being published), and I think it might be helpful to mention it earlier in the article (e.g. in an abstract if there was one).
