### 2020-06-01 khinsen/rescience-ten-year-challenge-paper-4#1 @khinsen
The definition of `guix-command` definitely relies on Org-Mode. The `#+NAME` is the equivalent of an alias in a shell script, but a bit more versatile. So to re-run my analysis using exactly the same versions of everything, you could do
```
alias guix-command=guix time-machine --commit=f76c16d220e6c349441c08bf25a5197037490fa5 --
```
assuming that I haven't overlooked some subtlety of shell aliases (which I haven't used for ages).
