### 2020-06-02 khinsen/rescience-ten-year-challenge-paper-4#1 @khinsen
I have rewritten the first paragraph to make sure that the version rendered by GitHub is not misleading. Now let's hope that GitHub won't change their rendering process and mess something up again!

Lesson learned: for reproducibility, don't rely on tools that evolve outside of your control. Pretty much impossible, but worth keeping in mind nevertheless.
