+++ article/content.tex
@@ -18,7 +18,7 @@ The \href{https://github.com/khinsen/rescience-ten-year-challenge-paper-4}{GitHu
 
 \subsection*{Finding the code}
 
-The main part of the code, HYDROLIB, is still available from the CPC Program Library \cite{CPCInternationalProgram2016} under the catalogue id \texttt{ADBK}, exactly as it was deposited in~1995. An updated version has been available on GitHub since~2015 \cite{HinsenHYDROLIBsourcecode2019}. The update does not modify the Fortran code, but only the compilation and installation scripts, which have in particular been adapted to the GNU~Fortran compiler. The user guide has also been revised to make it compatible with modern LaTeX installations (a minor change to three lines of the preamble).
+The main part of the code, HYDROLIB, is still available from the CPC Program Library \cite{CPCInternationalProgram2016} under the catalogue id \texttt{ADBK}, exactly as it was deposited in~1995. An updated version has been available on GitHub since~2015 \cite{HYDROLIB}. The update does not modify the Fortran code, but only the compilation and installation scripts, which have in particular been adapted to the GNU~Fortran compiler. The user guide has also been revised to make it compatible with modern LaTeX installations (a minor change to three lines of the preamble).
 
 The search for the Fortran programs that compute the sedimentation velocities of the nine conglomerates was more difficult and ultimately failed. Of the four computers that I used at the time, only one has survived: my personal Atari~TT. It is still in working condition, even though it tends to crash after about 30 minutes of operation, and has a copy of HYDROLIB on its disk, but not the programs for the nine conglomerates, most of which would have exceeded the computer's resources. The backup tapes from the DECstation have been lost during the intercontinental moves of my postdoc years.
 
