+++ article/bibliography.bib
@@ -72,11 +72,13 @@
   number = {2-3}
 }
 
-@misc{HinsenHYDROLIBsourcecode2019,
-  title = {{{HYDROLIB}} Source Code},
+@software{HYDROLIB,
+  title = {{HYDROLIB}},
   author = {Hinsen, Konrad},
   year = {2019},
-  howpublished = {https://github.com/khinsen/HYDROLIB}
+  swhid = {swh:1:dir:44d2058da36431234e4d6607f5ff001323472511;origin=https://github.com/khinsen/HYDROLIB},
+  license = {BSD},
+  url = {https://github.com/khinsen/HYDROLIB},
 }
 
 @article{HinsenRpStructuralflexibility2020,
