+++ article/article.tex
@@ -1,5 +1,14 @@
 \documentclass{rescience}
 
+% For software citations
+\usepackage{software-biblatex}
+\ExecuteBibliographyOptions{
+          halid=true,
+          swhid=true,
+          swlabels=true,
+          vcs=true,
+          license=true}
+
 \begin{document}
 \input{header.tex}
 \input{content.tex}
