### 2020-06-02 30a92cd
Fix typo in bibliography
