+++ article/bibliography.bib
@@ -14,7 +14,7 @@
 
 @article{CichockiFrictionmobilitymany1994,
   title = {Friction and Mobility of Many Spheres in {{Stokes}} Flow},
-  author = {Cichocki, B. and Felderhof, B. U. and Hinsen, K. and Wajnryb, E. and B\{l}awzdziewicz, J.},
+  author = {Cichocki, B. and Felderhof, B. U. and Hinsen, K. and Wajnryb, E. and B{\l}awzdziewicz, J.},
   year = {1994},
   month = mar,
   volume = {100},
