+++ article/content.tex
@@ -1,8 +1,8 @@
 \section*{Introduction}
 
-The article that is the subject of this reproduction attempt \cite{CichockiFrictionmobilitymany1994} describes the validation (by comparison to experiments) of a new method for computing hydrodynamic interactions in colloidal suspensions \cite{CichockiFrictionmobilitymany1994} and its implementation as a published Fortran~77 library called HYDROLIB \cite{HinsenHYDROLIBlibraryevaluation1995}. The latter is the first scientific software I ever published, at a time when this was a well-established possibility (the journal ``Computer Physics Communications'' was launched as early as 1969), but not yet common practice. The motivation for publishing this library was to make the method available to other researchers for inspection and use. Reproducibility was something that neither myself nor anyone else in the lab had ever heard of. The code specific to the validation paper was therefore not published.
+The article that is the subject of this reproduction attempt \cite{CichockiFrictionmobilitymany1994} describes the validation (by comparison to experiments) of a new method for computing hydrodynamic interactions in colloidal suspensions \cite{CichockiFrictionmobilitymany1994} and its implementation as a published Fortran~77 library called HYDROLIB \cite{HinsenHYDROLIBlibraryevaluation1995}. The latter is the first scientific software I ever published, at a time when this was a well-established possibility (the journal ``Computer Physics Communications'' was launched as early as 1969), but not yet common practice. The motivation for publishing this library was to make the method available to other researchers for inspection and use. Reproducibility was something that neither myself nor anyone else in the lab had ever heard of. The code specific to the validation paper was therefore not published. This turned out to be the major obstacle to reproducing the results of the original paper, as I was unable to find a copy of that project-specific code in my collection of old computers and backups.
 
-The validation consisted of computing the sedimentation velocities of several assemblies of spheres for which experimental measurements were available \cite{LassoStokesdraghollow1986}. These nine assemblies were regularly shaped conglomerates of two to 167 plastic spheres glued together. The description of the conglomerates provided by the experimenters was detailed enough to permit a computational reconstruction. The code written for the validation did exactly that: define the configurations of spheres for each of the nine conglomerates, then call a few subroutines from HYDROLIB to compute the sedimentation velocity.
+The validation of the new method consisted of computing the sedimentation velocities of several assemblies of spheres for which experimental measurements were available \cite{LassoStokesdraghollow1986}. These nine assemblies were regularly shaped conglomerates of two to 167 plastic spheres glued together. The description of the conglomerates provided by the experimenters was detailed enough to permit a computational reconstruction. The code written for the validation did exactly that: define the configurations of spheres for each of the nine conglomerates, then call a few subroutines from HYDROLIB to compute the sedimentation velocity.
 
 \section*{Historical context}
 
