+++ article/metadata.tex
@@ -5,13 +5,13 @@
 \def \codeSWH{}
 \def \dataURL{}
 \def \dataDOI{}
-\def \editorNAME{}
+\def \editorNAME{Tiziano Zito}
 \def \editorORCID{}
-\def \reviewerINAME{}
-\def \reviewerIORCID{}
+\def \reviewerINAME{Roman Yurchak}
+\def \reviewerIORCID{0000-0002-2565-4444}
 \def \reviewerIINAME{}
 \def \reviewerIIORCID{}
-\def \dateRECEIVED{}
+\def \dateRECEIVED{04 March 2020}
 \def \dateACCEPTED{}
 \def \datePUBLISHED{}
 \def \articleTITLE{[¬Rp] Stokes drag on conglomerates of spheres}
