### 2020-06-02 #20 @khinsen
Thanks @rth for this review!

Comments and actions taken:

1. I have rewritten the introduction to my code execution notes in order to make them clear even when viewed on GitHub (which strips away much of the Org-Mode markup). Let's hope that GitHub doesn't break this in an update to its rendered.

2. This is the first time I publish a computational protocol using Guix, so I am still experimenting with the best way to present everything. The reason for using a specific Guix environment for each individual command is the best possible documentation of the dependencies. For example, it's easy to see why `sed` is needed (for patching the installation script), and that the real computation does not depend on it. It is of course possible to use a single environment containing everything.

3. Leaving the environment explicitly is never required when working from Org-Mode, so it's not something I thought much about. I wouldn't do reproducible computations in an interactive shell anyway, because I would lose the output of the session. I have never used Guix in any other way than from Org-Mode, so I cannot say which other tools work well for this.

4. Thanks for the reminder about Software Heritage. Now that citing code has become easy with the recently published [biblatex-software](https://ctan.org/pkg/biblatex-software} package, I have updated my reference to the HYDROLIB library to point to SWH.

5. The typo in `summary.org` is fixed.

6. I have added a sentence about the nature of the failure to the first paragraph of the introduction.
