### 2020-06-06 #20 @otizonaizit
@khinsen : thank you for the revision! It looks good to me.
@rth : do you have additional comments?
