### 2020-06-09 #20 @rth
Thanks for addressing comments @khinsen ! Looks good to me. I don't have any other comments @otizonaizit .
