### 2020-06-10 #20 @otizonaizit
Great, so @khinsen the paper is hereby accepted! I'll go about publishing it in the next two days. I'll generate a PR against your repo for updating the metadata before publication.
