+++ article/metadata.yaml
@@ -95,14 +95,14 @@ contributors:
 # This information will be provided by the editor
 dates:
   - received: March 4, 2020
-  - accepted: 
-  - published: 
+  - accepted: June 10, 2020
+  - published: June 11, 2020
 
 # This information will be provided by the editor
 article:
-  - number: 
-  - doi:    
-  - url:    
+  - number: 7
+  - doi: 10.5281/zenodo.3889694
+  - url: https://zenodo.org/record/3889694/files/article.pdf
 
 # This information will be provided by the editor
 journal:
