+++ article/metadata.yaml
@@ -79,7 +79,7 @@ language: Fortran77
 
 # For example, the URL of the GitHub issue where review actually occured
 review: 
-  - url: 
+  - url: https://github.com/ReScience/submissions/issues/20 
 
 contributors:
   - name: Tiziano Zito
