### 2020-06-11 52361c1
Added SWH identifier
