### 2020-06-11 02640e7
Metadata in LaTeX
