+++ article/metadata.tex
@@ -2,7 +2,7 @@
 
 \def \codeURL{https://github.com/khinsen/rescience-ten-year-challenge-paper-4}
 \def \codeDOI{}
-\def \codeSWH{}
+\def \codeSWH{swh:1:dir:195b7890e84d3bcdc1fdbbbc351c63ac69a64ba0}
 \def \dataURL{}
 \def \dataDOI{}
 \def \editorNAME{Tiziano Zito}
@@ -12,14 +12,14 @@
 \def \reviewerIINAME{}
 \def \reviewerIIORCID{}
 \def \dateRECEIVED{04 March 2020}
-\def \dateACCEPTED{}
-\def \datePUBLISHED{}
+\def \dateACCEPTED{10 June 2020}
+\def \datePUBLISHED{11 June 2020}
 \def \articleTITLE{[¬Rp] Stokes drag on conglomerates of spheres}
 \def \articleTYPE{Reproduction}
 \def \articleDOMAIN{Fluid dynamics}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
 \def \articleYEAR{2020}
-\def \reviewURL{}
+\def \reviewURL{https://github.com/ReScience/submissions/issues/20}
 \def \articleABSTRACT{Reproduction attempt of B. Cichocki \& K. Hinsen, Stokes drag on conglomerates of spheres, Phys. Fluids 7:286 (1995). A reproduction turned out to be impossible because important parts of the code could not be retrieved. However, the surviving code still works without modification.}
 \def \replicationCITE{B. Cichocki, K. Hinsen, Stokes drag on conglomerates of spheres, Phys. Fluids 7:286 (1995)}
 \def \replicationBIB{CichockiStokesdragconglomerates1995}
@@ -31,8 +31,8 @@
 \def \journalNAME{ReScience C}
 \def \journalVOLUME{6}
 \def \journalISSUE{1}
-\def \articleNUMBER{}
-\def \articleDOI{}
+\def \articleNUMBER{7}
+\def \articleDOI{10.5281/zenodo.3889694}
 \def \authorsFULL{Konrad Hinsen}
 \def \authorsABBRV{K. Hinsen}
 \def \authorsSHORT{Hinsen}
