### 2020-06-11 #20 @otizonaizit
Publication is done: https://zenodo.org/record/3889694/files/article.pdf
Now waiting for the update to the website to make it official :-)
