### 2020-06-15 #20 @rougier
@otizonaizit Did you add the bibtex to the [website bibfile](https://github.com/ReScience/rescience.github.io/blob/sources/_bibliography/published.bib)?
