### 2020-06-15 #20 @otizonaizit
Yes, but there are open questions about it and we are waiting for your comments: https://github.com/ReScience/rescience.github.io/pull/81
