### 2020-08-19 #20 @khinsen
@otizonaizit Since this paper is now published, can we close this issue?
